Support a 2nd directory within the root dir
Add support for an extra directory with a known name, "rainbows".
Update nextFile/previousFile to skip over it (and other directories).
Add 2 macros for switching to playing videos in the rainbows directory
instead of the root directory (one for doing so only while the final
button of the macro is held, and one for switching indefinitely).  Add
another macro for returning to videos in the root directory.

When switching directories, preserve the state of which file we were in,
what our position was within the file, and what file is due to be opened
next.